机试题_晓建/Django_xx/DjangoHello/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
idna==2.8
pytz==2019.3
PyYAML==5.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.8
